feat #169 : 챌린지 생성 결과 화면 이동

+++ android/data/src/main/java/com/sixkids/data/repository/challenge/remote/ChallengeRemoteDataSource.kt
@@ -16,5 +16,5 @@ interface ChallengeRemoteDataSource {
         reword: Int,
         minCount: Int,
         groups: List<GroupSimple>
-    ): Long
+    ): Int
 }
