+++ android/domain/src/main/java/com/sixkids/domain/repository/ChallengeRepository.kt
@@ -24,5 +24,5 @@ interface ChallengeRepository {
         reword: Int,
         minCount: Int,
         groups: List<GroupSimple>
-    ): Long
+    ): Int
 }
